+++ regression/reference/generic-cfi/wrapfgeneric.f
@@ -172,7 +172,6 @@ module generic_mod
     ! Statement: c_function_native_scalar
     ! ----------------------------------------
     ! Argument:  const int * values +dimension(..)
-    ! Attrs:     +assumed-rank
     ! Statement: c_in_native_*
     ! ----------------------------------------
     ! Argument:  int nvalues
@@ -194,7 +193,6 @@ module generic_mod
     ! Statement: f_function_native_scalar
     ! ----------------------------------------
     ! Argument:  const int * values +dimension(..)
-    ! Attrs:     +assumed-rank
     ! Statement: f_in_native_*_cfi
     ! ----------------------------------------
     ! Argument:  int nvalues
@@ -869,7 +867,6 @@ contains
     ! Statement: f_function_native_scalar
     ! ----------------------------------------
     ! Argument:  const int * values +dimension(..)
-    ! Attrs:     +assumed-rank
     ! Statement: f_in_native_*_cfi
     ! ----------------------------------------
     ! Argument:  int nvalues